// frontend/src/core/api/generated/model/markRequest.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Kızılkan Şantiye API
 * OpenAPI spec version: 1.0
 */
import type { MarkRequestStatus } from './markRequestStatus'

export interface MarkRequest {
  status: MarkRequestStatus
  /** @nullable */
  overtimeHours?: number | null
  /**
   * @minLength 0
   * @maxLength 200
   * @nullable
   */
  note?: string | null
}